Move the PageParserContract into Contracts
