+++ src/Contracts/AbstractPageParser.php
@@ -3,7 +3,6 @@
 namespace Hyde\Framework\Contracts;
 
 use Hyde\Framework\Actions\ValidatesExistence;
-use Hyde\Framework\PageParserContract;
 
 /**
  * Abstract base class for all page parsers.
